refactored line graph into separate component, useEffec called once in DataVis then passed thru props to graphs
